requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.21
celery==3.1.19
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
joblib==0.9.3
kombu==3.0.33
MarkupSafe==0.23
nltk==3.1
numpy==1.10.1
pytz==2015.7
PyYAML==3.11
requests==2.8.1
scipy==0.16.1
six==1.10.0
Werkzeug==0.11.2
wheel==0.24.0
pyparsing==2.0.6
python-dateutil==2.4.2